chore rename type files
