--- src/types/public-types.ts
@@ -1,3 +1,0 @@
-export type * from './common-types.js';
-export type * from './createOracle-types.js';
-export type * from './updateOracle-types.js';
