+++ src/utils/getWeb3Provider.ts
@@ -3,7 +3,7 @@ import { getDefaultProvider } from '../config/config.js';
 import {
   Web3ReadOnlyProvider,
   Web3SignerProvider,
-} from '../types/common-types.js';
+} from '../types/common.js';
 
 export const getWeb3Provider = (privateKey: string): Web3SignerProvider =>
   getSignerFromPrivateKey('bellecour', privateKey);
